replay: extract to a separate module
