+++ src/meson.build
@@ -607,6 +607,8 @@ common_sources = [
   'trx/game/phase/phase_stats.c',
   'trx/game/photo_mode.c',
   'trx/game/random.c',
+  'trx/game/replay/test_recorder.c',
+  'trx/game/replay/test_replay.c',
   'trx/game/rooms/common.c',
   'trx/game/rooms/draw.c',
   'trx/game/rooms/floor_data.c',
@@ -634,8 +636,6 @@ common_sources = [
   'trx/game/stats/common.c',
   'trx/game/stats/init.c',
   'trx/game/stats/scan.c',
-  'trx/game/test_recorder.c',
-  'trx/game/test_replay.c',
   'trx/game/ui/common.c',
   'trx/game/ui/dialogs/base_passport.c',
   'trx/game/ui/dialogs/color_editor.c',
